requirements.txt
customtkinter==5.2.1
ping3==4.0.3
dnspython==2.4.0
speedtest-cli==2.1.3
python-whois==0.8.0
